+++ instrumentation/opentelemetry-instrumentation-asgi/src/opentelemetry/instrumentation/asgi/__init__.py
@@ -180,7 +180,7 @@ _ClientRequestHookT = typing.Optional[typing.Callable[[Span, dict], None]]
 _ClientResponseHookT = typing.Optional[typing.Callable[[Span, dict], None]]
 
 
-class ASGIGetter(Getter):
+class ASGIGetter(Getter[dict]):
     def get(
         self, carrier: dict, key: str
     ) -> typing.Optional[typing.List[str]]:
@@ -216,7 +216,7 @@ class ASGIGetter(Getter):
 asgi_getter = ASGIGetter()
 
 
-class ASGISetter(Setter):
+class ASGISetter(Setter[dict]):
     def set(
         self, carrier: dict, key: str, value: str
     ) -> None:  # pylint: disable=no-self-use