+++ propagator/opentelemetry-propagator-ot-trace/src/opentelemetry/propagators/ot_trace/__init__.py
@@ -53,7 +53,7 @@ class OTTracePropagator(TextMapPropagator):
         self,
         carrier: CarrierT,
         context: Optional[Context] = None,
-        getter: Getter = default_getter,
+        getter: Getter[CarrierT] = default_getter,
     ) -> Context:
         if context is None:
             context = Context()
@@ -112,7 +112,7 @@ class OTTracePropagator(TextMapPropagator):
         self,
         carrier: CarrierT,
         context: Optional[Context] = None,
-        setter: Setter = default_setter,
+        setter: Setter[CarrierT] = default_setter,
     ) -> None:
 
         span_context = get_current_span(context).get_span_context()